requirements.txt.txt
Flask 
google-generativeai 
gunicorn